+++ tycho-test/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod execution;
 pub mod rpc_tools;
-pub mod stream_processor;
 pub mod validation;
 pub use rpc_tools::RPCTools;